+++ Breakout/Vector2.h
@@ -13,8 +13,8 @@ public:
     Vector2 &operator+(Vector2);
 	Vector2 &operator-(Vector2);
 
-	Vector2 &operator*(double);
-	Vector2 &operator/(double);
+	Vector2 &operator*(float);
+	Vector2 &operator/(float);
 
 };
